1. Refactoring: Extract TransferCaseDocumentPublishService, remove midEvent handling 2. Fix unit tests (WIP)
