+++ src/test/java/uk/gov/hmcts/cmc/claimstore/services/ccd/callbacks/caseworker/transfercase/TransferCaseDocumentPublishServiceTest.java
@@ -1,0 +1,14 @@
+package uk.gov.hmcts.cmc.claimstore.services.ccd.callbacks.caseworker.transfercase;
+
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.fail;
+
+class TransferCaseDocumentPublishServiceTest {
+
+    @Test
+    void shouldPublishNoticesOfTransferToCase() {
+        // TODO
+        fail();
+    }
+}
