+++ src/test/java/uk/gov/hmcts/cmc/claimstore/services/ccd/callbacks/caseworker/transfercase/TransferCaseDocumentServiceTest.java
@@ -58,8 +58,6 @@ class TransferCaseDocumentServiceTest {
     @Test
     void shouldAttachNoticeOfTransferForDefendant() {
 
-        when(ccdCase.getDraftLetterDoc()).thenReturn(noticeForDefendant);
-
         transferCaseDocumentService.attachNoticeOfTransferForDefendant(ccdCase, noticeForDefendant, AUTHORISATION);
 
         verify(generalLetterService).attachGeneralLetterToCase(
