+++ src/test/java/uk/gov/hmcts/cmc/claimstore/services/ccd/callbacks/caseworker/transfercase/TransferCaseLetterSenderTest.java
@@ -48,7 +48,6 @@ class TransferCaseLetterSenderTest {
     void shouldSendNoticeOfTransferForDefendant() {
 
         CCDDocument noticeForDefendant = mock(CCDDocument.class);
-        when(ccdCase.getDraftLetterDoc()).thenReturn(noticeForDefendant);
 
         transferCaseLetterSender.sendNoticeOfTransferForDefendant(AUTHORISATION,
             noticeForDefendant, claim);
